third_party/blink/web_tests/external/wpt/css/css-values/minmax-length-serialize-expected.txt

This is a testharness.js-based test.
FAIL 'min(1px)' as a specified value should serialize as 'calc(1px)'. assert_equals: 'min(1px)' and 'calc(1px)' should serialize the same in specified values. expected "calc(1px)" but got "min(1px)"
PASS 'min(1px)' as a computed value should serialize as '1px'.
FAIL 'min(1in)' as a specified value should serialize as 'calc(96px)'. assert_equals: 'min(1in)' and 'calc(96px)' should serialize the same in specified values. expected "calc(96px)" but got "min(1in)"
PASS 'min(1in)' as a computed value should serialize as '96px'.
FAIL 'max(1px)' as a specified value should serialize as 'calc(1px)'. assert_equals: 'max(1px)' and 'calc(1px)' should serialize the same in specified values. expected "calc(1px)" but got "max(1px)"
PASS 'max(1px)' as a computed value should serialize as '1px'.
FAIL 'max(1in)' as a specified value should serialize as 'calc(96px)'. assert_equals: 'max(1in)' and 'calc(96px)' should serialize the same in specified values. expected "calc(96px)" but got "max(1in)"
PASS 'max(1in)' as a computed value should serialize as '96px'.
FAIL 'min(1PX)' as a specified value should serialize as 'calc(1px)'. assert_equals: 'min(1PX)' and 'calc(1px)' should serialize the same in specified values. expected "calc(1px)" but got "min(1px)"
PASS 'min(1PX)' as a computed value should serialize as '1px'.
FAIL 'min(50px, 1in + 1px)' as a specified value should serialize as 'calc(50px)'. assert_equals: 'min(50px, 1in + 1px)' and 'calc(50px)' should serialize the same in specified values. expected "calc(50px)" but got "min(50px, 97px)"
PASS 'min(50px, 1in + 1px)' as a computed value should serialize as '50px'.
FAIL 'max(50px, 1in + 1px)' as a specified value should serialize as 'calc(97px)'. assert_equals: 'max(50px, 1in + 1px)' and 'calc(97px)' should serialize the same in specified values. expected "calc(97px)" but got "max(50px, 97px)"
PASS 'max(50px, 1in + 1px)' as a computed value should serialize as '97px'.
FAIL 'calc(1px + min(1in, 100px))' as a specified value should serialize as 'calc(97px)'. assert_equals: 'calc(1px + min(1in, 100px))' and 'calc(97px)' should serialize the same in specified values. expected "calc(97px)" but got "calc(1px + min(1in, 100px))"
PASS 'calc(1px + min(1in, 100px))' as a computed value should serialize as '97px'.
FAIL 'calc(1px + max(1in, 100px))' as a specified value should serialize as 'calc(101px)'. assert_equals: 'calc(1px + max(1in, 100px))' and 'calc(101px)' should serialize the same in specified values. expected "calc(101px)" but got "calc(1px + max(1in, 100px))"
PASS 'calc(1px + max(1in, 100px))' as a computed value should serialize as '101px'.
PASS 'min(1px, 1em)' as a specified value should serialize as 'min(1px, 1em)'.
PASS 'min(1px, 1em)' as a computed value should serialize as '1px'.
FAIL 'calc(min(1px, 1in) + max(100px + 1em, 10px + 1in) + 1px)' as a specified value should serialize as 'calc(2px + max(1em + 100px, 106px))'. assert_equals: 'calc(min(1px, 1in) + max(100px + 1em, 10px + 1in) + 1px)' and 'calc(2px + max(1em + 100px, 106px))' should serialize the same in specified values. expected "calc(2px + max(1em + 100px, 106px))" but got "calc(min(1px, 1in) + max(100px + 1em, 106px) + 1px)"
PASS 'calc(min(1px, 1in) + max(100px + 1em, 10px + 1in) + 1px)' as a computed value should serialize as '118px'.
Harness: the test ran to completion.
